Move start_game tests to utilities
